+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## `v0.0.62`
+
+### Improvements
+
+- Txs are now broadcasted directly via the user-specified RPC instead of via the wallet (closes <https://github.com/coinhall/cosmes/issues/17>)
+- Update Station extension to use Keplr's API interface
+
 ## `v0.0.60`
 
 ### Features